requirements.txt
aiohttp==3.0.9
aiohttp-jinja2==0.17.0
aiohttpdemo-polls==0.0.1
aiomysql==0.0.12
aiopg==0.13.2
aliyun-python-sdk-core-v3==2.8.6
aliyun-python-sdk-dysmsapi==1.0.0
async-timeout==2.0.1
attrs==17.4.0
certifi==2018.1.18
chardet==3.0.4
diff-match-patch==20121119
Django==2.0.3
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.0
django-reversion==2.0.13
et-xmlfile==1.0.1
future==0.16.0
get==0.0.39
httplib2==0.11.3
idna==2.6
idna-ssl==1.0.1
jdcal==1.4
Jinja2==2.10
MarkupSafe==1.0
multidict==4.1.0
mysqlclient==1.3.12
odfpy==1.3.6
openpyxl==2.5.2
Pillow==5.1.0
post==0.0.26
psycopg2==2.7.4
public==0.0.65
PyMySQL==0.8.0
pytz==2018.3
PyYAML==3.12
query-string==0.0.28
request==0.0.26
requests==2.18.4
six==1.11.0
South==1.0.2
SQLAlchemy==1.2.5
tablib==0.12.1
trafaret==1.1.1
trafaret-config==1.0.1
unicodecsv==0.14.1
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
yarl==1.1.1
you-get==0.4.1060
